requirements.txt
fastapi
uvicorn
redis
openai
pyyaml
rich 